fix(refactor): Refactored class components to function components
