+++ src/App.js
@@ -9,7 +9,7 @@ import React, { useEffect, useCallback } from 'react'
 import './App.css'
 import LoadScripts from './components/load-scripts'
 import AsyncLoad from './services/async-load'
-import SelectServerButton from './components/servers/select-server-button'
+import SelectServerButton from './components/app-body/servers/select-server-button'
 import Footer from './components/footer'
 import NavMenu from './components/nav-menu'
 import useAppState from './hooks/state'
