+++ src/components/app-body/index.js
@@ -13,7 +13,7 @@ import React from 'react'
 import GetBalance from './balance'
 import Placeholder2 from './placeholder2'
 import Placeholder3 from './placeholder3'
-import ServerSelectView from '../servers/select-server-view'
+import ServerSelectView from './servers/select-server-view'
 
 function AppBody (props) {
   // Dependency injection through props
