+++ src/components/app-body/placeholder2.js
@@ -3,22 +3,18 @@
 */
 
 // Global npm libraries
-import React from 'react'
+import React, { useEffect } from 'react'
 
-class Placeholder2 extends React.Component {
-  constructor (props) {
-    super(props)
-
+function Placeholder2 (props) {
+  useEffect(() => {
     console.log('Placeholder 2 loaded.')
-  }
+  }, [])
 
-  render () {
-    return (
-      <>
-        <p style={{ padding: '25px' }}>This is placeholder View #2</p>
-      </>
-    )
-  }
+  return (
+    <>
+      <p style={{ padding: '25px' }}>This is placeholder View #2</p>
+    </>
+  )
 }
 
 export default Placeholder2
